make glfw work on mac
